Py 312 pytoml (#4)
* Refactor project structure to use pyproject.toml

This commit represents a significant refactor of the project structure. It replaces the setup.py and requirements.txt with a pyproject.toml which contains all the necessary information about the project and its dependencies. The python version in Github workflow is also updated from "3.10" to "3.12". Changes also include modifications to test and linting configurations for better compatibility with the refactored structure.

* Update Bandit and Bumper configurations

This commit refines the Bandit and Bumper tool configurations in the "pyproject.toml" file, moving specific settings for each tool into the respective tool sections. In the 'Bandit' section, the targets, exclusion directories, and skipped tests are specified. In the 'Bumper' section, the exclusion files, version files, repository, and report are clearly stated. Made corresponding changes to the GitHub workflow file, replacing hardcoded options with the settings defined in "pyproject.toml".

Signed-off-by: Steven K <steven@pointcircle.com>


---------

Signed-off-by: Steven K <steven@pointcircle.com>